refactor: move permission actions seed to entity
