+++ src/api/data/permission/permission.entity.ts
@@ -1,4 +1,14 @@
 import { Permission } from '@prisma/client';
 import { PrismaBaseEntity } from '../prismaBaseEntity.js';
 
+export const actions = {
+  post: ['readPost', 'createPost', 'updatePost', 'trashPost', 'publishPost', 'archivePost'],
+  project: ['readProject', 'updateProject'],
+  role: ['readRole', 'createRole', 'updateRole', 'deleteRole'],
+  user: ['readUser', 'updateUser', 'deleteUser'],
+  invitation: ['inviteUser'],
+  review: ['readOwnReview', 'readAllReview', 'createReview', 'approveReview', 'closeReview'],
+  apiKey: ['createApiKey', 'readApiKey', 'updateApiKey', 'deleteApiKey'],
+};
+
 export class PermissionEntity extends PrismaBaseEntity<Permission> {}
